Generalized the infinite scrollable list
